Finished logic for rest of commands

+++ HMI_ECU/src/eeprom/eeprom_driver.h
@@ -4,6 +4,6 @@
 
 int compare_passwords(const uint8_t *password);
 
-void change_password(const int *new_password);
+void change_password(const uint8_t *new_password);
 
 #endif //EEPROM_DRIVER_H
